MMSB-4 Adding Docker Added Docker into the project and using it to build and run the application successfully
Every thing is OK if we run this application in local machine. However, when we run it through Docker container, there is an error in connecting to DB. Here is the detail information about this error:
HTTP request:
http://localhost:8080/user/user/1
Error response:
There was an unexpected error (type=Internal Server Error, status=500).
Could not open JDBC Connection for transaction; nested exception is com.mysql.jdbc.exceptions.jdbc4.CommunicationsException: Communications link failure The last packet sent successfully to the server was 0 milliseconds ago. The driver has not received any packets from the server.